Don't set CHOICESPATH (Thomas Leonard). From my mailing list post:
It shouldn't really be setting it to anything. After much hunting through the
svn logs, it seems that this behaviour was introduced in revision 3497 (in the C
version) when I got it to get the shared MIME database through the old Zero
Install /uri filesystem, and has been preserved ever since even though that code
was later removed!


git-svn-id: svn://svn.code.sf.net/p/rox/svn/trunk/ROX-Session@4915 66de3db3-b00d-0410-b41b-f4738ad19bea